requirements.txt
certifi==2019.11.28+sx
chardet==3.0.4
DateTime==4.3
idna==2.8
numpy==1.18.1
pandas==1.0.1
pypredict==1.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
zope.interface==4.7.1
